+++ index.css
@@ -50,12 +50,12 @@ p {
   border: none;
   font-weight: 600;
   border-left: 2px solid black;
+  color: black;
+  cursor: pointer;
 }
 
 .navitem:first-child {
   border-left: none;
-  color: blue;
-  cursor: auto;
 }
 
 .navitem:hover {
@@ -174,8 +174,7 @@ button {
   display: flex;
   border: 2px solid blue;
   padding: 5px;
-  margin: 30px 0;
   position: fixed;
   bottom: 20px;
-  width: 85%;
+  width: calc(100% - 70px);
 }
