+++ index.html
@@ -29,7 +29,7 @@
         <form id="form">
             <input class="formelement" type="text" id="title" placeholder="Title" required>
             <input class="formelement" type="text" id="author" placeholder="Author" required>
-            <button class="formelement" id="button" type="submit">Add</button>
+            <button class="formelement" id="button" type="button">Add</button>
         </form>
     </section>
     <section id="contactsection">
@@ -45,6 +45,6 @@
     <footer id="footer">
         Copyright
     </footer>
-    <script src="./index.js"></script>
+    <script type="module" src="./index.js"></script>
 </body>
 </html>
